+++ src/common/types.d.ts
@@ -1,4 +1,11 @@
 declare module "common-types" {
+  export type ReactRenderer = {
+    id: number;
+    name: string;
+    version: string;
+    channelId: `events:${number}`;
+  };
+
   export type FiberType = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
   export type FiberRootMode = 0 | 1 | 2;