+++ src/publisher/index.ts
@@ -1,10 +1,12 @@
 import { getHost } from "rempl";
 import config from "./config";
 import { installReactDevtoolsHook } from "./react-devtools-hook";
-import { recordEvent } from "./rempl-publisher";
+import { publishReactInstance } from "./rempl-publisher";
 import { attach } from "./react-integration";
 
-installReactDevtoolsHook(window, renderer => attach(renderer, recordEvent));
+installReactDevtoolsHook(window, (id, renderer) =>
+  attach(renderer, publishReactInstance(id, renderer))
+);
 
 if (config.inpage) {
   getHost().activate();