+++ src/publisher/rempl-publisher.ts
@@ -1,9 +1,11 @@
 import { createPublisher } from "rempl";
 import debounce from "lodash.debounce";
-import { RecordEventHandler, Message } from "./types";
+import { ReactRenderer } from "common-types";
+import { RecordEventHandler, Message, ReactInternals } from "./types";
 import { ToolId } from "../common/constants";
 import config from "./config";
 import { resolveSourceLoc } from "./utils/resolveSourceLoc";
+import { getRendererInfo } from "./utils/renderer-info";
 
 let eventIdSeed = 0;
 const { openSourceLoc } = config;
@@ -20,7 +22,8 @@ declare let __SUBSCRIBER_SRC__: string;
 
 export const publisher = createPublisher(ToolId, (settings, callback) => {
   if (__DEV__) {
-    const { origin } = new URL(import.meta.url);
+    // const { origin } = new URL(import.meta.url);
+    const origin = "http://localhost:3000";
 
     fetch(`${origin}/subscriber.js`)
       .then(res => res.text())
@@ -30,59 +33,84 @@ export const publisher = createPublisher(ToolId, (settings, callback) => {
   }
 });
 
-const eventLogChannel = publisher.ns("tree-changes");
-eventLogChannel.provide("getEvents", (offset, count) => {
-  if (isNaN(offset) || isNaN(count)) {
-    return [];
-  }
+const reactInstances: ReactRenderer[] = [];
+const reactInstancesChannel = publisher.ns("react-renderers");
 
-  publishEventsDebounced.flush();
+reactInstancesChannel.publish([]);
 
-  const start = Math.max(0, Math.floor(offset));
-  let end =
-    start + Math.min(Math.max(0, Math.floor(count)), events.length - start);
+export function publishReactInstance(id: number, renderer: ReactInternals) {
+  const channelId = `events:${id}` as `events:${number}`;
 
-  if (end > start) {
-    const { commitId } = events[end - 1];
-    for (; end < events.length; end++) {
-      if (events[end].commitId !== commitId) {
-        break;
-      }
-    }
-  }
-
-  return events.slice(start, end);
-});
-const getEventsState = () => ({
-  count: events.length,
-});
-eventLogChannel.provide("getEventsState", () => {
-  publishEventsDebounced.flush();
-  return getEventsState();
-});
-const publishEventsDebounced = debounce(
-  () => eventLogChannel.publish(getEventsState()),
-  50,
-  { maxWait: 50 }
-);
-export const recordEvent: RecordEventHandler = payload => {
-  const id = eventIdSeed++;
-
-  events.push({
-    id,
-    // timestamp: Math.trunc(getTimestamp()),
-    ...payload,
+  reactInstances.push({
+    ...getRendererInfo(id, renderer),
+    channelId,
   });
 
-  publishEventsDebounced();
+  const eventLogChannel = publisher.ns(channelId);
+  const getEventsState = () => ({
+    count: events.length,
+  });
+  const publishEventsDebounced = debounce(
+    () => eventLogChannel.publish(getEventsState()),
+    50,
+    { maxWait: 50 }
+  );
+  const recordEvent: RecordEventHandler = payload => {
+    const id = eventIdSeed++;
 
-  return id;
-};
+    events.push({
+      id,
+      // timestamp: Math.trunc(getTimestamp()),
+      ...payload,
+    });
+
+    publishEventsDebounced();
+
+    return id;
+  };
+
+  reactInstancesChannel.publish(reactInstances);
+  eventLogChannel.publish(getEventsState());
+  eventLogChannel.provide({
+    getEventsState() {
+      publishEventsDebounced.flush();
+
+      return getEventsState();
+    },
+    getEvents(offset, count) {
+      if (isNaN(offset) || isNaN(count)) {
+        return [];
+      }
+
+      publishEventsDebounced.flush();
+
+      const start = Math.max(0, Math.floor(offset));
+      let end =
+        start + Math.min(Math.max(0, Math.floor(count)), events.length - start);
+
+      if (end > start) {
+        const { commitId } = events[end - 1];
+        for (; end < events.length; end++) {
+          if (events[end].commitId !== commitId) {
+            break;
+          }
+        }
+      }
+
+      return events.slice(start, end);
+    },
+  });
+
+  return recordEvent;
+}
 
 publisher.ns("open-source-settings").publish(openSourceLoc || null);
-
 publisher.provide("resolve-source-locations", locations =>
   Promise.all(locations.map(resolveSourceLoc)).then(result =>
     result.map((resolved, idx) => ({ loc: locations[idx], resolved }))
   )
 );
+
+// console.log("!!!!");
+// import { connectPublisherWs } from 'rempl';
+// connectPublisherWs("http://localhost:8177/");