+++ src/ui/components/misc/WaitingForReady.tsx
@@ -3,21 +3,15 @@ import { useEventsContext } from "../../utils/events";
 import { useFiberChildren } from "../../utils/fiber-maps";
 
 export default function WaitingForReady() {
-  const [visible, setVisible] = React.useState(false);
   const children = useFiberChildren(0);
   const { loadedEventsCount, totalEventsCount } = useEventsContext();
-
-  // Use effect to trigger a transition
-  React.useEffect(() => {
-    setVisible(true);
-  }, []);
 
   if (children.length > 0) {
     return null;
   }
 
   return (
-    <div className={"waiting-for-ready" + (visible ? " visible" : "")}>
+    <div className="waiting-for-ready">
       {totalEventsCount > 0
         ? loadedEventsCount === totalEventsCount
           ? "Rendering..."